Yet another backend for hidl-gen, this one generates a C-compatible header file
containing those enum types annotated in the package like so:

@export
enum Foo {
    ...
};

Optionally, the name to be used for the type declaration in the header file
can be different from that used in the .hal interface description by specifying

@export(name="foo_t")
enum Foo {
    ...
};

Finally, overriding the name to be empty, i.e.

@export(name="")
enum Foo {
    ...
};

will cause the generator to emit an anonymous enum.

Bug: 31800672
Change-Id: Idffb2c1700af1c7fd312941d80c3373add8ae558
Test: make