Wrap compound native_handle_t* in hidl_pointer.
When transferring a handle in a compound type,
we need to use hidl_pointer<> to ensure that we
allocate enough space for the pointer. But we want
to hide the hidl_pointer<> type from client/server
code as much as we can, so we only use this type
when necessary (eg when the handle is embedded in
a struct).

Bug: 32089785
Test: builds
Change-Id: I30c9aa9f1348fd5e72cd6d1e71d9ac3b4bd8ea8e