Fix java-client generation of HIDL code
hidl_vec<T> and hidl_string was recently changed to use uint32_t for mSize
field instead of size_t. See ag/1640383
As a result Java clients started to crash because the size of hidl_vec and
hidl_string was reduced to 16 bytes and Java gen code was not adjusted.

Test: our java client is no longer crashing

Change-Id: I506ecd61277e3a4df96a10180c46ea924a42235d
Fix: b/33038659