Wait for binderized service if VINTF promises it.
We should wait for the binderized service if the VINTF says it is on the
device. This way the framework will not open a service in passthrough
that is actually a hwbinder service.

Here is a chart of the interaction between getStub and VINTF transport
(to make the behavior of getService as it stand clear and because it's
not explained elsewhere):

              |        getStub
VINTF         | true      | false
----------------------------------------
Passthrough   | (1) PT    | (4) [HB ->] PT
-----------------------------------------
HwBinder      | (2) PT    | (5) HB wait
-----------------------------------------
Unspecified   | (3) PT    | (6) HB -> PT

1). get the service as specified in passthrough
2). get passthrough service (this should only ever be called from
registerPassthroughServiceImplementation).
3). default to pre-VINTF behavior (for now)
4). look for binderized service first (in case ENABLE_TREBLE is true)
then fallback to passthrough as specified in the VINTF. This behavior
will have to be changed.
5). (this CL) Block and wait for hwbinder service since VINTF promises
it will be there.
6). default to pre-VINTF behavior (for now)

Bug: 31458381
Test: hidl_test, nfc/wifi/fingerprint/light/etc work on internal master
with ENABLE_TREBLE on or off
Change-Id: Ic722d1facce257759486f2090d613c04c360c0d1