Do proper cleanup of static block.
The static initializer of each HAL library adds
entries to global maps in a different static library,
but it never deletes them. This causes problems if
the HAL library is ever unloaded later, because the
map will be pointing to entries that no longer exist.

To fix this, switch to using __attribute__((constructor))
and ((destructor)), so we will clean up when we need to.

Bug: 34950358
Test: drm_hidl_test, hidl_test
Change-Id: If81b9712a651a07390fcc6641d86a77bb7a03f7c