getService: respect vintf passthrough
I had getService continue to load the hwbinder service even if the vintf
specified passthrough because in order to make development with
ENABLE_TREBLE work with development of the vintf. However, now that a
sufficiently large number of people know about the vintf and it has been
tested, it's time to take this out.

Bug: 34274385
Test: works with master
Change-Id: I12767e904f8a498de2041c95154aa0436308b937