Add tryGetService.
Required for cases where a hal is being used but we want to fail in
cases where the remote HAL is down. This is also required when someone
wants to get a handle on a hidl service from the only hwbinder thread in
a pool.

Test: hidl_test
Bug: 36780273
Merged-In: I6544f534ef2dff0773c645cc9953de6d1340d01d
Change-Id: I6544f534ef2dff0773c645cc9953de6d1340d01d