Omit HIDL test libraries from the VNDK.
Add -t option to hidl-gen to emit omit vndk enabled when generating
Android.bp. This is to filter-out the shared libs of HALs for testing
from the VNDK directory (system/lib/vndk) to save storage.

update-makefile-helper.sh is also updated to set the -t option when
there exists .hidl_for_test file in the directory for an HAL package.

Bug: 64776708
Test: no tests.vendor*.so file in /system/lib[64]/vndk on 2017 pixel
devices
Test: hidl_test

Change-Id: Idb3a14bdecbda43309db8ef862baab59ff8ba647