Build HIDL interfaces for testing with cc_test_library
Add -t option to hidl-gen to emit cc_test_library when generating
Android.bp. This is to filter-out the shared libs of HALs for testing
from the VNDK directory (system/lib/vndk) to save storage.

update-makefile-helper.sh is also updated to set the -t option when
there exists .hidl_for_test file in the directory for an HAL package.

Bug: 64776708
Test: no tests.vendor*.so file in /system/lib[64]/vndk on 2017 pixel
devices

Change-Id: If7f46edd75b149afdd119b3ff9a0a21fc0a55646