Add C++ forward declaration check
C++ prohibits forward declaration of an inner type.
At this point we take it over and prohibit such forward references.

The formal rule is that if type A references in its declaration type B
then B is parent of A or B’s parent is parent of A
(stack from global scope to A contains the exact parent of A).

Bug: 31827278

Test: mma
Test: test in follow-up CL moving lookups
Change-Id: Ia163fb76bce243c826b992202849108acd53f2c1