Add hidl-gen-host_test.
Note the primary difference between this and utils_test is that
these tests are host only whereas utils_test is testing the part
of hidl-gen that is available both on host and on device.

This only contains tests for Coordinator and Location because they
have been modified recently. More tests will be added in the future.

Also updated run_all_host_tests.sh to accept additional host tests.

Test: run_all_host_tests
Change-Id: Ie5a0727157ea109f2e6d408c1d5acc84517e9587