Fix bug in enforceMinorVersionUprevs.
Every minor version uprev should extend the top-level
interfaces from the previous minor version. In hidl-gen,
we check that at least one interface is extended. Here,
there was a problem where if the minor version upgrade
contains an interface which is not in the previous package,
then this was counted as an extended interface.

The exact problem was that this statement:
lastFQName.getPackageAndVersion() == prevPackage.getPackageAndVersion()

considered an interface extended even if !lastFQNameExists. Since
lastFQName was always constructed from prevPackage, this statement
is always true.

Test: problematic interface fails
Test: hidl_version_test
Change-Id: I41983ad5c44eae1903123d49a9106509e4005f7b