Add implementors notes to -Landroidbp-impl
There is occasionally confusion both internally and externally
as what the correct thing is to do to create a hal. Here I have
added additional implementors notes explaining how to convert
the boilerplate that is generated automatically into an actual
implementation of a hal.

Note that a '-impl' passthrough implementation is still generated.
Because there is no concept of the "top-level" interface, we are
unable to generate a 'service.cpp' file because we don't know which
interface is actually correct to regester. Further, in many cases
multiple hals will be served or they are served from existing
daemons.

Test: manual + run_all_host_tests
Change-Id: I81573b78a89c9d4e147195bfcfb89c5e7dc8c1d3