Add graphics common@1.1 to the list of sp hals.
This package had its makefile edited manually, but
I'm updating it here so that hidl-gen will generate
it properly.

The list is kept here because this is purely a policy
thing and there is no way to distinguish what should
or shouldn't be allowed to be same process without
out-of-source-code knowledge.

Bug: N/A
Test: Makefiles are updated consistently.
Change-Id: Icd27c6e33ddba331fa2f082c42671db46bfa67d2