Unify Java + C++ flag definitions.
Right now, we have duplicated definitions of flags in
C++ and Java that have to be kept in sync. Moving this here
so that that is no longer a requirement.

This will also allow us to add new flags in the future after
we start emitting transaction errors when flags are

Bug: 72877008
Test: N/A
Change-Id: I17b1ff55d09200f09e02edfdc2f1619bb121b3da