Unify Java + C++ flag definitions. (2/2)
Second CL for this b/c branching.

Right now, we have duplicated definitions of flags in
C++ and Java that have to be kept in sync. Moving this here
so that that is no longer a requirement.

This will also allow us to add new flags in the future after
we start emitting transaction errors when flags are

Bug: 72877008
Test: N/A
Change-Id: I4b3d8fdc1a54f06e7abd189f12d5701ea525db30