Dump verbose output for all -L targets.
Before this, printing defined but unreferenced types
was restricted to when '-v' was set on a target that
generated sources for a target. Now it works on all
of them (to increase visibility). This makes deps
easier to implement because it removes logic that is
attached to a function which encodes file structure
in its control flow (and in the future will be moved
into an object).

Bug: 69526027
Test: hidl's run_all_host_tests.sh
Change-Id: Iadcaf9c03730fd4251fe7c87d3f878c1f95bbcb7