Fix link-type warnings on espresso-* prebult libs
The prebuilt libs are all marked with LOCAL_SDK_VERSION so that they
don't trigger link-type check warning when they are used from a library
or an app built with SDK.

For example, espresso-core is built with SDK and using
espresso-core-nodep and espresso-idling-resource-nodep, which causes
link-type check warning. This warning will soon be error.

Bug: 69899800
Test: mma -j under this directory. The build is successufl and does not
show link-type check warning other than for tagsoup (which is currently
being handled).

Change-Id: I1a9d71405ab71901d454bb2fab651353ea9ea671