Fix incremental builds.
Turns out, hidl-gen is trying to read some files that
don't exist. This happens because types files recognize
import ...::Iterator; // for an ambiguous case

And try to open the correspond HAL file as an interface
and then try to open types.hal to find the type.

This CL moves the deps addition until after the existance
of the file is determined.

In the future, two solutions are possible:
- require "::types.Iterator" to refer to a type in HIDL
- build system keeps track of which files are deleted so
  that we can declare a dependency on files we *would*
  read

Test: m && m (second output does nothing).
Fixes: 73359909
Change-Id: If25e15608ccb7ef06c3a1d57902bd8d949588f03