Import fails if parsing fails.
Previously, hidl-gen doesn't distinguish between
failed parse due to parsing errors or due to the
file being missing. This means that any of a number
of times when a file was imported, it could fail
to parse, but the parsing of the file that referenced
it would continue on happily.

A case to hidl_error_test was also added which fails before
this patch and succeeds after.

Change-Id: Ie51a56eff6ef35c7ae5c513252a3a7e6581addb1
Fixes: 64773282
Test: hidl's run_all_host_tests.sh